+++ packages/gestalt/src/Masonry/defaultLayout.test.ts
@@ -246,6 +246,44 @@ describe.each([undefined, getColumnSpanConfig])('default layout tests', (_getCol
       { top: 0, left: 600, width: 100, height: 100 },
     ]);
   });
+
+  test('correctly positions items with no height', () => {
+    const measurementStore = new MeasurementStore<Record<any, any>, number>();
+    const positionCache = new MeasurementStore<Record<any, any>, Position>();
+    const items: ReadonlyArray<Item> = [
+      { 'name': 'Pin 0', 'height': 100 },
+      { 'name': 'Pin 1', 'height': 120 },
+      { 'name': 'Pin 2', 'height': 0 },
+      { 'name': 'Pin 3', 'height': 100 },
+      { 'name': 'Pin 4', 'height': 100 },
+      { 'name': 'Pin 5', 'height': 120 },
+      { 'name': 'Pin 6', 'height': 80 },
+      { 'name': 'Pin 7', 'height': 100 },
+    ];
+    items.forEach((item: any) => {
+      measurementStore.set(item, item.height);
+    });
+
+    const layout = defaultLayout({
+      align: 'end',
+      measurementCache: measurementStore,
+      positionCache,
+      layout: 'basic',
+      minCols: 2,
+      rawItemCount: items.length,
+      width: 1000,
+      _getColumnSpanConfig,
+    });
+
+    const positions = layout(items);
+    const pin2Position = positions[2];
+    const pin3Position = positions[3];
+
+    expect(pin2Position.height).toBe(0);
+    expect(pin2Position.top).toBe(0);
+    expect(pin3Position.top).toBe(0);
+    expect(pin2Position.left).toBe(pin3Position.left);
+  });
 });
 
 describe('loadingStateItems', () => {