+++ packages/gestalt/src/index.ts
@@ -76,7 +76,9 @@ import TapArea from './TapArea';
 import TapAreaLink from './TapAreaLink';
 import Text from './Text';
 import TextArea from './TextArea';
+import TextCompact from './TextCompact';
 import TextField from './TextField';
+import TextUI from './TextUI';
 import TileData from './TileData';
 import Toast from './Toast';
 import Tooltip from './Tooltip';
@@ -165,7 +167,9 @@ export {
   TapAreaLink,
   Text,
   TextArea,
+  TextCompact,
   TextField,
+  TextUI,
   TileData,
   Toast,
   Tooltip,